FEAT(#48): EXTEND LAYOUT IN DONATIONS LIST VIEW
