+++ main/templates/layouts/master.html
@@ -27,8 +27,8 @@
     {% load static %}
     <link rel="stylesheet" type="text/css" href="{% static 'styles/master.css' %}">
     <link rel="stylesheet" type="text/css" href="{% static 'styles/master_sidebar.css' %}">
-
-    <link rel="stylesheet" type="text/css" href="{% static 'styles/donation.css' %}">
+    {% block extrahead %} {% endblock %}
+    
 
 </head>
