+++ main/templates/layouts/master_sidebar.html
@@ -1,4 +1,4 @@
-{% extends 'master.html' %}
+{% extends 'layouts/master.html' %}
 
 {% block contenido %}
 <div class="p-4">
