+++ main/urls.py
@@ -1,7 +1,8 @@
-from django.urls import path
+from django.urls import path, include
 from . import views
+import donation.urls as donation_urls
 
 urlpatterns = [
-    path('', views.index, name='home'),
-    path('donations',views.donations_list, name='donations_list')
+    path('', views.index, name='index'),
+    path('donations/', include(donation_urls), name='donations'),
 ]
